Renamed component from LoginForm; map userLogin Action to props; if user.isLoggedIn and no errors from login action, render /quiz
